requirements.txt
beautifulsoup4~=4.12.2
lxml~=4.9.2
requests~=2.28.2
python-dotenv~=1.0.0
icecream~=2.1.3
schedule~=1.2.0
boto3~=1.26.114